+++ src/main/java/net/linkednest/common/CommonConstants.java
