+++ src/main/java/net/linkednest/common/entity/Authority.java
