+++ src/main/java/net/linkednest/common/filter/JwtAuthenticationFilter.java
@@ -1,11 +1,11 @@
-package net.linkednest.www.filter;
+package net.linkednest.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import net.linkednest.www.security.JwtProvider;
+import net.linkednest.security.JwtProvider;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
