+++ src/main/java/net/linkednest/common/security/CustomUserDetails.java
@@ -1,6 +1,6 @@
-package net.linkednest.www.security;
+package net.linkednest.security;
 
-import net.linkednest.www.entity.User;
+import net.linkednest.common.entity.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
