+++ src/main/java/net/linkednest/common/utils/CommonUtil.java
